refactor(workspace): move durable record beside resolver
Separate workspace metadata from store mutation choreography.
